Add quiet limit order short circuit. Refactor tests to use pkg version (#33)
The proposed contract changes here are extremely small. The goal is to
prevent limit order instructions from rolling back the full transaction
if the user has insufficient funds.

The new test checks this logic against multiple different inputs.

Changes:
- Create an `amount` function for the token checker. This is pulled
directly from Anchor:
https://github.com/coral-xyz/anchor/blob/master/spl/src/token.rs#L534
- Create logic in the new order processor to quietly exit the function
in the case of insufficient funds:
    1) Check that the user specified that the tx should fail silently
2) Check the available quote lots and base lots (free + token balance)
3) For either bid or ask, make sure that the trader has sufficient
available balance to perform the new operation
    4) If the funds are insufficient, exit
- Update the test file to use a far more recent version of phoenix-sdk
(this is the majority of the changes in this PR). Note that this doesn't
affect the contract at all
- Write a new test case to validate that the short circuiting logic
works as intended
- Add a new field to OrderPacket to enable the silent failing case.
- Update the deserializer to maintain backwards compatibility