Address minor ordering issue in PlaceMultipleOrders (#39)
The logic in PlaceMultipleOrders sorts both bids and asks from lowest
price to highest price. When combined with the feature to silently fail
on insufficient funds, this results in slightly unexpected behavior
where the wider (less aggressive) levels of the bid orders are placed
before the tighter levels. If the user runs out of quote collateral in
this process, the wide bid levels will be displayed but the tight levels
will be silently omitted.

The opposite but desired effect is true for the ask orders.